migrate AuthFlowResult conversion to EventData into extension method
