+++ src/AzureAuth.Test/CommandAadTest.cs
@@ -445,72 +445,6 @@ invalid_key = ""this is not a valid alias key""
             subject.TokenFetcherOptions.Should().BeEquivalentTo(expected);
         }
 
-        /// <summary>
-        /// Test to generate event data from a null authflow result.
-        /// </summary>
-        [Test]
-        public void TestGenerateEvent_FromNullAuthResult()
-        {
-            AuthFlowResult authFlowResult = null;
-            var subject = this.serviceProvider.GetService<CommandAad>();
-
-            // Act
-            var eventData = subject.AuthFlowEventData(authFlowResult);
-
-            // Assert
-            eventData.Should().BeNull();
-        }
-
-        /// <summary>
-        /// Test to generate event data from an authflow result with null token result and null errors.
-        /// </summary>
-        [Test]
-        public void TestGenerateEvent_From_AuthFlowResult_With_Null_TokenResult_Null_Errors()
-        {
-            AuthFlowResult authFlowResult = new AuthFlowResult(null, null, "AuthFlowName");
-            var subject = this.serviceProvider.GetService<CommandAad>();
-
-            // Act
-            var eventData = subject.AuthFlowEventData(authFlowResult);
-
-            // Assert
-            eventData.Properties.Should().NotContainKey("msal_correlation_ids");
-            eventData.Properties.Should().NotContainKey("error_messages");
-            eventData.Measures.Should().NotContainKey("token_validity_hours");
-            eventData.Properties.Should().NotContainKey("silent");
-
-            eventData.Properties.Should().Contain("authflow", "AuthFlowName");
-            eventData.Properties.Should().Contain("success", "False");
-            eventData.Measures.Should().ContainKey("duration_milliseconds");
-        }
-
-        /// <summary>
-        /// Test to generate event data from an authflow result with null token result and some errors.
-        /// </summary>
-        [Test]
-        public void TestGenerateEvent_From_AuthFlowResult_With_Errors_And_Null_TokenResult()
-        {
-            var errors = new[]
-            {
-                new Exception("Exception 1."),
-            };
-
-            AuthFlowResult authFlowResult = new AuthFlowResult(null, errors, "AuthFlowName");
-            var subject = this.serviceProvider.GetService<CommandAad>();
-
-            // Act
-            var eventData = subject.AuthFlowEventData(authFlowResult);
-
-            // Assert
-            eventData.Properties.Should().NotContainKey("msal_correlation_ids");
-            eventData.Measures.Should().NotContainKey("token_validity_minutes");
-            eventData.Properties.Should().NotContainKey("silent");
-            eventData.Properties.Should().Contain("authflow", "AuthFlowName");
-            eventData.Properties.Should().Contain("success", "False");
-            eventData.Properties.Should().Contain("error_messages", "[{\"Message\":\"Exception 1.\",\"InnerException\":null,\"ExceptionType\":\"System.Exception\",\"AADErrorCode\":null,\"CorrelationId\":null}]");
-            eventData.Measures.Should().ContainKey("duration_milliseconds");
-        }
-
         [Test]
         public void TestSendEvent_From_AuthFlowResult_With_Errors_And_Null_TokenResult()
         {
@@ -543,73 +477,6 @@ invalid_key = ""this is not a valid alias key""
 
             // Assert
             this.telemetryServiceMock.VerifyAll();
-        }
-
-        /// <summary>
-        /// Test to generate event data from an authflow result with token result and msal errors.
-        /// </summary>
-        [Test]
-        public void TestGenerateEvent_From_AuthFlowResult_With_MsalErrors_And_TokenResult()
-        {
-            // TODO
-            var correlationID1 = Guid.NewGuid().ToString();
-            var msalServiceException = new MsalServiceException("1", "An MSAL Service Exception message");
-            msalServiceException.CorrelationId = correlationID1;
-
-            var msalUIRequiredException = new MsalUiRequiredException("2", "An MSAL UI Required Exception message");
-            msalUIRequiredException.CorrelationId = null;
-
-            var errors = new[]
-            {
-                msalServiceException,
-                msalUIRequiredException,
-            };
-
-            var tokenResultCorrelationID = Guid.NewGuid();
-            var tokenResult = new TokenResult(new JsonWebToken(FakeToken), tokenResultCorrelationID);
-
-            AuthFlowResult authFlowResult = new AuthFlowResult(tokenResult, errors, "AuthFlowName");
-            var subject = this.serviceProvider.GetService<CommandAad>();
-
-            // Act
-            var eventData = subject.AuthFlowEventData(authFlowResult);
-
-            // Assert
-            eventData.Properties.Should().Contain("authflow", "AuthFlowName");
-            eventData.Properties.Should().Contain("success", "True");
-            eventData.Properties.Should().Contain("msal_correlation_id", $"{tokenResultCorrelationID}");
-            eventData.Properties.Should().Contain("silent", "False");
-            eventData.Properties.Should().Contain("error_messages", "[{\"Message\":\"An MSAL Service Exception message\",\"InnerException\":null,\"ExceptionType\":\"Microsoft.Identity.Client.MsalServiceException\",\"AADErrorCode\":\"AADSTS1\",\"CorrelationId\":\"" + $"{correlationID1}" + "\"},{\"Message\":\"An MSAL UI Required Exception message\",\"InnerException\":null,\"ExceptionType\":\"Microsoft.Identity.Client.MsalUiRequiredException\",\"AADErrorCode\":\"AADSTS2\",\"CorrelationId\":null}]");
-            eventData.Measures.Should().ContainKey("token_validity_minutes");
-            eventData.Measures.Should().ContainKey("duration_milliseconds");
-        }
-
-        /// <summary>
-        /// Test to generate event data from an authflow result with token result and no errors.
-        /// </summary>
-        [Test]
-        public void TestGenerateEvent_From_AuthFlowResult_With_TokenResult_And_Null_Errors()
-        {
-            // TODO
-            var tokenResultCorrelationID = Guid.NewGuid();
-            var tokenResult = new TokenResult(new JsonWebToken(FakeToken), tokenResultCorrelationID);
-
-            AuthFlowResult authFlowResult = new AuthFlowResult(tokenResult, null, "AuthFlowName");
-            var subject = this.serviceProvider.GetService<CommandAad>();
-
-            var expectedCorrelationIDs = $"{tokenResultCorrelationID}";
-
-            // Act
-            var eventData = subject.AuthFlowEventData(authFlowResult);
-
-            // Assert
-            eventData.Properties.Should().NotContainKey("error_messages");
-            eventData.Properties.Should().Contain("authflow", "AuthFlowName");
-            eventData.Properties.Should().Contain("success", "True");
-            eventData.Properties.Should().Contain("msal_correlation_id", expectedCorrelationIDs);
-            eventData.Properties.Should().Contain("silent", "False");
-            eventData.Measures.Should().ContainKey("token_validity_minutes");
-            eventData.Measures.Should().ContainKey("duration_milliseconds");
         }
 
         [TestCase("1", true)]
